Condense header and footer code
