+++ app/webpacker/components/CompetitionForm/Footer.js
@@ -1,29 +1,127 @@
-import React from 'react';
-import { Message } from 'semantic-ui-react';
+import { Button } from 'semantic-ui-react';
+import React, { useMemo } from 'react';
+import I18n from '../../lib/i18n';
 import { useStore } from '../../lib/providers/StoreProvider';
-import { useFormContext } from '../wca/FormBuilder/provider/FormObjectProvider';
-import ConfirmationActions, { CreateOrUpdateButton } from './ConfirmationActions';
+import useLoadedData from '../../lib/hooks/useLoadedData';
+import {
+  competitionConfirmationDataUrl,
+  competitionUrl,
+  confirmCompetitionUrl,
+  homepageUrl,
+} from '../../lib/requests/routes.js.erb';
+import Loading from '../Requests/Loading';
+import ConfirmProvider, { useConfirm } from '../../lib/providers/ConfirmProvider';
+import useSaveAction from '../../lib/hooks/useSaveAction';
+import { useFormCommitAction, useFormUpdateAction } from '../wca/FormBuilder/EditForm';
+import { useFormErrorHandler, useFormInitialObject } from '../wca/FormBuilder/provider/FormObjectProvider';
 
-export default function Footer({
-  saveObject,
+function ConfirmButton({
+  competitionId,
+  data,
+  sync,
 }) {
-  const { isPersisted } = useStore();
-  const { unsavedChanges } = useFormContext();
+  const { canConfirm } = data;
 
-  if (isPersisted && !unsavedChanges) {
-    return (
-      <ConfirmationActions saveObject={saveObject} />
-    );
-  }
+  const onError = useFormErrorHandler();
+
+  const { save } = useSaveAction();
+  const confirm = useConfirm();
+
+  const updateFormObject = useFormUpdateAction();
+  const commitFormObject = useFormCommitAction();
+
+  const confirmCompetition = () => {
+    confirm({
+      content: I18n.t('competitions.competition_form.submit_confirm'),
+    }).then(() => {
+      save(confirmCompetitionUrl(competitionId), null, () => {
+        sync();
+
+        // mark the competition as announced and commit immediately.
+        // (we do not want the announce button to trigger the "there are unsaved changes" alert)
+        updateFormObject('isConfirmed', true, ['admin']);
+        commitFormObject();
+      }, {
+        body: null,
+        method: 'PUT',
+      }, onError);
+    });
+  };
+
+  if (!canConfirm) return null;
 
   return (
-    <>
-      {unsavedChanges && (
-        <Message info>
-          You have unsaved changes. Please save the competition before taking any other action.
-        </Message>
-      )}
-      <CreateOrUpdateButton saveObject={saveObject} />
-    </>
+    <Button
+      positive
+      onClick={confirmCompetition}
+    >
+      {I18n.t('competitions.competition_form.submit_confirm_value')}
+    </Button>
+  );
+}
+
+function DeleteButton({
+  competitionId,
+  data,
+}) {
+  const { cannotDeleteReason } = data;
+
+  const { save } = useSaveAction();
+  const confirm = useConfirm();
+
+  const deleteCompetition = () => {
+    confirm({
+      content: I18n.t('competitions.competition_form.submit_delete'),
+    }).then(() => {
+      save(competitionUrl(competitionId), null, () => {
+        window.location.replace(homepageUrl);
+      }, {
+        body: null,
+        method: 'DELETE',
+      });
+    });
+  };
+
+  if (cannotDeleteReason) return null;
+
+  return (
+    <Button
+      negative
+      onClick={deleteCompetition}
+    >
+      {I18n.t('competitions.competition_form.submit_delete_value')}
+    </Button>
+  );
+}
+
+export default function Footer() {
+  const { isAdminView } = useStore();
+
+  const {
+    competitionId,
+    admin: { isConfirmed },
+  } = useFormInitialObject();
+
+  const dataUrl = useMemo(() => competitionConfirmationDataUrl(competitionId), [competitionId]);
+
+  const {
+    data,
+    loading,
+    sync,
+  } = useLoadedData(dataUrl);
+
+  if (loading) return <Loading />;
+
+  return (
+    <ConfirmProvider>
+      <Button.Group>
+        {!isAdminView && !isConfirmed && (
+          <ConfirmButton competitionId={competitionId} data={data} sync={sync} />
+        )}
+        {!isConfirmed && (
+          <DeleteButton competitionId={competitionId} data={data} />
+        )}
+      </Button.Group>
+    </ConfirmProvider>
   );
 }
